Menu updates
Made a much smoother transition into the game from the menu buttons.
Also, the mute button works now and remembers what it was set to the
last time you played.  If you could make the mute button look good Dan
that would be awesome.  it basically just has to be a white bar going
diagonally, but that's far too complicated for my artistic skills.