Changes to friendly spawning algorithm
Noticed a bug or two. I think they are fixed, but I noticed they always
like to spawn in the bottom right corner now....

Will double check later.

Also going to change the algorithm so that it always performs 100
searches, then uses the best one instead of doing 1-100 and picking one
that works.